[Created] สร้าง function คำนวนค่าไม่มีทุนทรัพย์ตามมาตรา 288(1)

+++ src/noncapital/noncapital.go
@@ -11,13 +11,13 @@ func CheckFeeTypeNonCapital(feeType string) bool {
 	return false
 }
 
-func InitFeePrice0() int {
+func InitFeePrice0() float64 {
 
-	return 0
+	return 0.00
 }
 
-func InitFeePrice200() int {
-	return 200
+func InitFeePrice200() float64 {
+	return 200.00
 }
 
 func CheckSectionType288_1(section string) bool {
